requirements.txt
PySide6
pyyaml
xmltodict
